+++ plugins/senspark/platform/android/FacebookProtocolUser.cpp
@@ -72,7 +72,7 @@ void FacebookProtocolUser::graphRequest(const std::string& graphPath, const FBPa
     PluginParam pathParam(graphPath.c_str());
     PluginParam paramsParam(params);
     CallbackWrapper *wrapper = new CallbackWrapper(callback);
-    PluginParam cbID((long)wrapper);
+    PluginParam cbID((long long)wrapper);
 
     callFuncWithParam("graphRequestWithParams", &pathParam, &paramsParam, &cbID, nullptr);
 }
@@ -83,7 +83,7 @@ void FacebookProtocolUser::api(const std::string &graphPath, HttpMethod method, 
     PluginParam _method((int) method);
     PluginParam paramsParam(params);
     CallbackWrapper *wrapper = new CallbackWrapper(callback);
-    PluginParam cbID((long)wrapper);
+    PluginParam cbID((long long)wrapper);
 
     callFuncWithParam("api", &pathParam, &_method, &paramsParam, &cbID, NULL);
 }
