+++ plugins/senspark/platform/android/GooglePlayProtocolSocial.cpp
@@ -40,7 +40,7 @@ void GooglePlayProtocolSocial::revealAchievement(const std::string& achievementI
     PluginParam idParam(achievementId.c_str());
 
     CallbackWrapper* wrapper = new CallbackWrapper(pcb);
-    PluginParam cbParam((long) wrapper);
+    PluginParam cbParam((long long) wrapper);
 
     callFuncWithParam("revealAchievement", &idParam, &cbParam, nullptr);
 }
@@ -49,7 +49,7 @@ void GooglePlayProtocolSocial::resetAchievement(const std::string& achievementId
     PluginParam idParam(achievementId.c_str());
 
     CallbackWrapper* wrapper = new CallbackWrapper(pcb);
-    PluginParam cbParam((long) wrapper);
+    PluginParam cbParam((long long) wrapper);
 
     callFuncWithParam("resetAchievement", &idParam, &cbParam, nullptr);
 }
@@ -57,7 +57,7 @@ void GooglePlayProtocolSocial::resetAchievement(const std::string& achievementId
 void GooglePlayProtocolSocial::resetAchievements(const GooglePlayProtocolSocial::SocialCallback& pcb) {
 
     CallbackWrapper* wrapper = new CallbackWrapper(pcb);
-    PluginParam cbParam((long) wrapper);
+    PluginParam cbParam((long long) wrapper);
 
 
     callFuncWithParam("resetAchievements", &cbParam, nullptr);
