+++ plugins/senspark/platform/ios/ParseProtocolBaaS.mm
@@ -31,7 +31,7 @@ std::string ParseProtocolBaaS::setUserInfo(const std::string& jsonChanges) {
 void ParseProtocolBaaS::saveUserInfo(BaaSCallback &cb) {
     CallbackWrapper *wrapper = new CallbackWrapper(cb);
     
-    PluginParam callbackParam((long) wrapper);
+    PluginParam callbackParam((long long) wrapper);
     
     callFuncWithParam("saveUserInfo", &callbackParam, nullptr);
 }
@@ -39,7 +39,7 @@ void ParseProtocolBaaS::saveUserInfo(BaaSCallback &cb) {
 void ParseProtocolBaaS::fetchUserInfo(BaaSCallback &cb) {
     CallbackWrapper *wrapper = new CallbackWrapper(cb);
     
-    PluginParam callbackParam((long) wrapper);
+    PluginParam callbackParam((long long) wrapper);
     
     callFuncWithParam("fetchUserInfo", &callbackParam, nullptr);
 }
@@ -56,14 +56,14 @@ std::string ParseProtocolBaaS::setInstallationInfo(const std::string& changes) {
 
 void ParseProtocolBaaS::saveInstallationInfo(BaaSCallback& cb) {
     CallbackWrapper *wrapper = new CallbackWrapper(cb);
-    PluginParam callbackParam((long) wrapper);
+    PluginParam callbackParam((long long) wrapper);
     callFuncWithParam("saveInstallationInfo", &callbackParam, nullptr);
 }
 
 void ParseProtocolBaaS::loginWithFacebookAccessToken(BaaSCallback& cb) {
     CallbackWrapper *wrapper = new CallbackWrapper(cb);
     
-    PluginParam callbackParam((long) wrapper);
+    PluginParam callbackParam((long long) wrapper);
     
     callFuncWithParam("loginWithFacebookAccessToken", &callbackParam, nullptr);
 }
@@ -87,7 +87,7 @@ std::string ParseProtocolBaaS::getSubscribedChannels() {
 void ParseProtocolBaaS::fetchConfigInBackground(BaaSCallback& cb) {
     CallbackWrapper *wrapper = new CallbackWrapper(cb);
     
-    PluginParam callbackParam((long) wrapper);
+    PluginParam callbackParam((long long) wrapper);
     
     callFuncWithParam("fetchConfigInBackground", &callbackParam, nullptr);
 }
